+++ javaSE/src/main/java/com/smartdevicelink/managers/SdlManager.java
@@ -36,6 +36,7 @@ import android.support.annotation.NonNull;
 import android.util.Log;
 import com.smartdevicelink.managers.file.FileManager;
 import com.smartdevicelink.managers.file.filetypes.SdlArtwork;
+import com.smartdevicelink.managers.lifecycle.LifecycleConfigurationUpdate;
 import com.smartdevicelink.managers.lifecycle.LifecycleManager;
 import com.smartdevicelink.managers.permission.PermissionManager;
 import com.smartdevicelink.managers.screen.ScreenManager;
@@ -43,20 +44,24 @@ import com.smartdevicelink.protocol.enums.FunctionID;
 import com.smartdevicelink.protocol.enums.SessionType;
 import com.smartdevicelink.proxy.RPCMessage;
 import com.smartdevicelink.proxy.RPCRequest;
+import com.smartdevicelink.proxy.RPCResponse;
 import com.smartdevicelink.proxy.SystemCapabilityManager;
 import com.smartdevicelink.proxy.interfaces.ISdl;
 import com.smartdevicelink.proxy.rpc.*;
 import com.smartdevicelink.proxy.rpc.enums.AppHMIType;
 import com.smartdevicelink.proxy.rpc.enums.Language;
+import com.smartdevicelink.proxy.rpc.enums.Result;
 import com.smartdevicelink.proxy.rpc.enums.SdlDisconnectedReason;
 import com.smartdevicelink.proxy.rpc.listeners.OnMultipleRequestListener;
 import com.smartdevicelink.proxy.rpc.listeners.OnRPCNotificationListener;
 import com.smartdevicelink.proxy.rpc.listeners.OnRPCRequestListener;
+import com.smartdevicelink.proxy.rpc.listeners.OnRPCResponseListener;
 import com.smartdevicelink.security.SdlSecurityBase;
 import com.smartdevicelink.transport.BaseTransportConfig;
 import com.smartdevicelink.transport.enums.TransportType;
 import com.smartdevicelink.util.DebugTool;
 import com.smartdevicelink.util.Version;
+import org.json.JSONException;
 
 import java.util.*;
 
@@ -215,7 +220,51 @@ public class SdlManager extends BaseSdlManager{
 
     @Override
     protected void checkLifecycleConfiguration(){
-		this.checkLifecycleConfigurationPrivate(managerListener);
+		Language actualLanguage =  lifecycleManager.getRegisterAppInterfaceResponse().getLanguage();
+		LifecycleConfigurationUpdate lcu = managerListener.managerShouldUpdateLifecycle(actualLanguage);
+
+		if (lcu != null){
+			// go through and change sdlManager properties that were changed via the LCU update
+			hmiLanguage = actualLanguage;
+
+			if (lcu.getAppName() != null){
+				appName = lcu.getAppName();
+			}
+
+			if (lcu.getShortAppName() != null){
+				shortAppName = lcu.getShortAppName();
+			}
+
+			if (lcu.getTtsName() != null){
+				ttsChunks = lcu.getTtsName();
+			}
+
+			if (lcu.getVoiceRecognitionCommandNames() != null){
+				vrSynonyms = lcu.getVoiceRecognitionCommandNames();
+			}
+
+			ChangeRegistration changeRegistration = new ChangeRegistration(actualLanguage,actualLanguage);
+			changeRegistration.setAppName(lcu.getAppName());
+			changeRegistration.setNgnMediaScreenAppName(lcu.getShortAppName());
+			changeRegistration.setTtsName(lcu.getTtsName());
+			changeRegistration.setVrSynonyms(lcu.getVoiceRecognitionCommandNames());
+			changeRegistration.setOnRPCResponseListener(new OnRPCResponseListener() {
+				@Override
+				public void onResponse(int correlationId, RPCResponse response) {
+					try {
+						Log.v(TAG, response.serializeJSON().toString());
+					} catch (JSONException e) {
+						e.printStackTrace();
+					}
+				}
+
+				@Override
+				public void onError(int correlationId, Result resultCode, String info){
+					Log.e(TAG, "Change Registration onError: "+ resultCode+ " | Info: "+ info);
+				}
+			});
+			_internalInterface.sendRPC(changeRegistration);
+		}
     }
 
 	@Override
